+++ lib/badge-cli.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const { PDFKitTextMeasurer } = require('./measure-text');
+const { PDFKitTextMeasurer } = require('./text-measurer');
 const { makeBadge } = require('./make-badge');
 const svg2img = require('./svg-to-img');
 const colorscheme = require('./colorscheme.json');
