+++ lib/make-badge-test-helpers.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const { PDFKitTextMeasurer } = require('./measure-text');
+const { PDFKitTextMeasurer } = require('./text-measurer');
 const { makeMakeBadgeFn } = require('./make-badge');
 
 module.exports = {
