+++ lib/text-measurer.js
