+++ lib/text-measurer.spec.js
