+++ server.js
@@ -15,7 +15,7 @@ const githubAuth = require('./lib/github-auth');
 const sysMonitor = require('./lib/sys/monitor');
 const log = require('./lib/log');
 const { makeMakeBadgeFn } = require('./lib/make-badge');
-const { QuickTextMeasurer } = require('./lib/measure-text');
+const { QuickTextMeasurer } = require('./lib/text-measurer');
 const serverSecrets = require('./lib/server-secrets');
 const suggest = require('./lib/suggest');
 const {licenseToColor} = require('./lib/licenses');
